+++ app/src/main/java/com/example/rasago/data/entity/CustomerEntity.kt
@@ -5,7 +5,7 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "customers")
 data class CustomerEntity(
-    @PrimaryKey(autoGenerate = true) val customerId: Long = 0,
+    @PrimaryKey(autoGenerate = true) val customerId: Int = 0,
     val name: String,
     val phone: String,
     val email: String,
